Fix default port handling
Fix the default handling for ports, particularly when re-using a builder for multiple registrations, where the scheme is changed without explicitly changing the port.
This was causing the default port for the first scheme to be carried through into registrations for subsequent schemes as creating a UriBuilder from another copies the Port value through, causing it to appear to have been explicitly set, rather than keeping the "use default for scheme" behaviour.