Log response matching
Log tests for matching of HTTP requests to responses.
To make this "nice", this introduces a concept of adding Ids to registrations when using the builders. If an Id is not set, an auto-generated one is created based on the interception's HTTP method, URI, status code and hash code is used instead.