Fix NullReferenceException
Fix NullReferenceException deregistering a builder if it used a custom match delegate. Now it will deregister as long as it hasn't been mutated (otherwise it is "lost" and can't be removed).
If the registration is "lost", then an InvalidOperationException is thrown instead to draw the caller's attention to the fact their code is not behaving as expected.
Resolves #361.